safety commit - organize helpers
