--- src/components/victory-chart/scale-helpers.js
@@ -1,8 +1,0 @@
-import Scale from "../../helpers/scale";
-
-module.exports = {
-  getScale(props, axisComponent, axis) {
-    return Scale.getScaleFromProps(props, axis) ||
-      axisComponent.type.getScale(axisComponent.props);
-  }
-};
